Provide an explicit second name in test
This test is simply to ensure that we get a useful warning from having
two columns with the same name. Having one of them empty is causing a
different error in some contexts, as there usually must be at least one
of id | name provided. So have an explicit value the second time around
to properly isolate what we're actually testing.