Do not produce Executive data
Previously, if there was an 'executive' or 'post' column, we would turn
that into a set of memberships of an executive body, separately from the
other memberships to the legislative body. This was unneeded complexity,
and all known users of this code needed to post-process the data to
handle this in odd ways. So we no longer support it.