No longer partitiion by a 'house'
Previously, if there was a 'house' or 'chamber' column, we would create
separate sub-legislature organizations, and make the memberships be of
those instead.

This is unneeded complexity, and all known users of this code needed to
post-process the data to handle this in odd ways. So we no longer
support it.